create store dir
